added gemfile
