+++ main.rb
@@ -1,24 +1,26 @@
 # main.rb
 require_relative 'alien'
+require_relative 'guest_list'
 require_relative 'solar_system'
 require 'colorize'
+require 'lolize/auto'
 puts ""
+
 def main
   earth = SolarSystem.new("Earth")
   venus = SolarSystem.new("Venus")
   mars = SolarSystem.new("Mars")
+  guest = SolarSystem.new("Guest List")
 
- guest = SolarSystem.new("Guest List")
-
-  zetas = Alien_Type.new('Zetas', 'gray', 5.97324244, 'Reticuli binary star system', 'Zetas are intellectually ahead of us by a few million years or so.')
   felines = Alien_Type.new('The Felines', 'gold, white, but it varies', 88973242, 'Sirius Star System', 'They exist from 3D on up to 12D!')
+  reptiles = Alien_Type.new('Reptiles', 'similar to human skin tone', 9, 'Pleiades Star Cluster', 'They communicate by means of light transfer.')
   pleiadians = Alien_Type.new('Pleiadians', 'similar to human skin tone', 9, 'Pleiades Star Cluster', 'They communicate by means of light transfer.')
+  zetas = Alien_Type.new('Zetas', 'gray', 5.97324244, 'Reticuli binary star system', 'Zetas are intellectually ahead of us by a few million years or so.')
 
-  # earth.add_alien(felines)
-  # venus.add_alien(pleiadians)
-  # mars.add_alien(zetas)
-  # guest.user_added_alien
-
+  earth.add_alien(felines)
+  earth.add_alien(reptiles)
+  venus.add_alien(pleiadians)
+  mars.add_alien(zetas)
 
   exit_loop = false
 
@@ -34,37 +36,28 @@ def main
     user_input = gets.chomp
 
     if user_input == "list aliens"
-      puts guest.list_user_aliens.colorize(:light_blue)
-      # puts venus.list_aliens.colorize(:light_magenta)
-      # puts mars.list_aliens.colorize(:light_red)
+      puts ""
+      puts earth.list_aliens.colorize(:light_green)
+      puts venus.list_aliens.colorize(:light_magenta)
+      puts mars.list_aliens.colorize(:light_red)
+      puts guest.list_user_aliens.colorize(:light_cyan)
       puts ""
     end
 
-    def add_alien1
+    def new_guest_alien
       puts ""
       puts "Let's add a new alien! Please provide the following info:"
       puts ""
       print "Alien name: " 
       name = gets.chomp
-      print "Alien color: "
-      color = gets.chomp
-      print "Alien home, star system or star cluster etc. "
-      home = gets.chomp
-      print "Alien's distance from the sun in light years: "
-      distance = gets.chomp
-      print "Fun fact about the alien: "
-      fun_fact = gets.chomp
-      puts ""
-    
-      brand_new_alien = Alien_Type.new(name, color, home, distance, fun_fact)
+     
+      brand_new_alien = Guest_Type.new(name)
       return brand_new_alien
     end
 
     if user_input == "add an alien"
-      brand_new_alien = add_alien1
+      brand_new_alien = new_guest_alien
       guest.user_added_alien(brand_new_alien)
-
-      
     end
 
     if user_input == "read alien details"
